Create module sale_by_packaging
The functions of this module were first added into sale_order_line_packaging_qty.

After sorting out and extracting the different functions in multiple modules in
the previous commits, this one gets it all together in a single module.
