--- product_packaging_can_be_sold/models/__init__.py
@@ -1,2 +1,0 @@
-from . import product_packaging
-from . import sale_order_line
