--- product_packaging_can_be_sold/models/product_packaging.py
@@ -1,9 +1,0 @@
-# Copyright 2020 Camptocamp SA
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl)
-from odoo import fields, models
-
-
-class ProductPackaging(models.Model):
-    _inherit = "product.packaging"
-
-    can_be_sold = fields.Boolean(string="Can be sold", default=True,)
