--- product_packaging_can_be_sold/readme/DESCRIPTION.rst
@@ -1,1 +1,0 @@
-This module allows to restrict product packaging that can be sold.
