--- product_packaging_can_be_sold/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_packaging_can_be_sold
