--- product_packaging_type_can_be_sold/__init__.py
@@ -1,1 +1,0 @@
-from . import models
