--- product_packaging_type_can_be_sold/readme/CONTRIBUTORS.rst
@@ -1,2 +1,0 @@
-* Thomas Nowicki <thomas.nowicki@camptocamp.com>
-* Akim Juillerat <akim.juillerat@camptocamp.com>
