--- product_packaging_type_can_be_sold/readme/DESCRIPTION.rst
@@ -1,3 +1,0 @@
-This module extends the functionality of module "Product Packaging Can Be Sold"
-to by moving the "can be sold" flag from the product packaging to its product
-packaging type.
