+++ sale_by_packaging/models/__init__.py
@@ -1,0 +1,5 @@
+from . import product_packaging
+from . import product_packaging_type
+from . import product_product
+from . import product_template
+from . import sale_order_line
