+++ sale_by_packaging/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Akim Juillerat <akim.juillerat@camptocamp.com>
+* Thomas Nowicki <thomas.nowicki@camptocamp.com>
