+++ sale_by_packaging/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module provides different configuration option to manage packagings on
+sale orders.
