+++ sale_by_packaging/tests/__init__.py
@@ -1,0 +1,2 @@
+from . import test_packaging_type_can_be_sold
+from . import test_sale_only_by_packaging
