+++ sale_by_packaging/views/product_template.xml
