--- sale_product_by_packaging_only/__init__.py
@@ -1,1 +1,0 @@
-from . import models
