--- sale_product_by_packaging_only/__manifest__.py
@@ -1,16 +1,0 @@
-# Copyright 2020 Camptocamp SA
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl)
-{
-    "name": "Sale Product By Packaging Only",
-    "summary": "Restrict sales from product without packaging",
-    "version": "13.0.1.1.0",
-    "development_status": "Alpha",
-    "category": "Warehouse Management",
-    "website": "https://github.com/OCA/sale-workflow",
-    "author": "Camptocamp, Odoo Community Association (OCA)",
-    "license": "AGPL-3",
-    "application": False,
-    "installable": True,
-    "depends": ["sale_stock"],
-    "data": ["views/product_template.xml"],
-}
