--- sale_product_by_packaging_only/models/__init__.py
@@ -1,3 +1,0 @@
-from . import product_product
-from . import product_template
-from . import sale_order_line
