--- sale_product_by_packaging_only/readme/CONTRIBUTORS.rst
@@ -1,2 +1,0 @@
-* Thomas Nowicki <thomas.nowicki@camptocamp.com>
-* Akim Juillerat <akim.juillerat@camptocamp.com>
