--- sale_product_by_packaging_only/readme/DESCRIPTION.rst
@@ -1,2 +1,0 @@
-This module restricts the possibility of selling products without having a
-packaging defined.
