--- sale_product_by_packaging_only/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_sale_only_by_packaging
