--- sale_product_packaging_restriction/__init__.py
@@ -1,1 +1,0 @@
-from . import models
