--- sale_product_packaging_restriction/__manifest__.py
@@ -1,17 +1,0 @@
-# Copyright 2020 Camptocamp SA
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl)
-{
-    "name": "Sale Product Packaging Restriction",
-    "summary": "Restrict sales of packagings",
-    "version": "13.0.1.1.0",
-    "development_status": "Alpha",
-    "category": "Warehouse Management",
-    "website": "https://github.com/OCA/sale-workflow",
-    "author": "Camptocamp, Odoo Community Association (OCA)",
-    "license": "AGPL-3",
-    "application": False,
-    "installable": True,
-    "auto-install": True,
-    "depends": ["product_packaging_can_be_sold", "sale_product_by_packaging_only"],
-    "data": [],
-}
