--- sale_product_packaging_restriction/models/__init__.py
@@ -1,2 +1,0 @@
-from . import product_product
-from . import product_template
