--- sale_product_packaging_restriction/readme/CONTRIBUTORS.rst
@@ -1,1 +1,0 @@
-* Akim Juillerat <akim.juillerat@camptocamp.com>
