--- sale_product_packaging_restriction/readme/DESCRIPTION.rst
@@ -1,2 +1,0 @@
-This module provides some glue between product_packaging_can_be_sold and
-sale_product_by_packaging_only modules.
