--- setup/product_packaging_type_can_be_sold/odoo/addons/product_packaging_type_can_be_sold
@@ -1,1 +1,0 @@
-../../../../product_packaging_type_can_be_sold
