+++ setup/sale_by_packaging/odoo/addons/sale_by_packaging
@@ -1,0 +1,1 @@
+../../../../sale_by_packaging
