--- setup/sale_product_by_packaging_only/odoo/addons/sale_product_by_packaging_only
@@ -1,1 +1,0 @@
-../../../../sale_product_by_packaging_only
