--- setup/sale_product_packaging_restriction/odoo/addons/sale_product_packaging_restriction
@@ -1,1 +1,0 @@
-../../../../sale_product_packaging_restriction
